+++ game/src/dlls/objects/423/423.c
@@ -293,7 +293,8 @@ void EdibleMushroom_updateBehavior(GameObject* obj, EdibleMushroomState* state, 
                         }
                         state->pickupMsgValue = 0;
                         state->pickupMsgDelay = 0.4f;
-                        ObjMsg_SendToObject(player, EDIBLE_MUSHROOM_MESSAGE_IN_RANGE, obj, (u32)&state->pickupMsgBitId);
+                        ObjMsg_SendToObject(player, EDIBLE_MUSHROOM_MESSAGE_IN_RANGE, obj,
+                                            (uintptr_t)&state->pickupMsgBitId);
                         bit = ObjAnim_ReadPlacementS16(&obj->anim, &(placement->gameBitId));
                         if (bit != -1) {
                             mainSetBits(bit, 1);
